setup node and typescript
