+++ package.json
@@ -1,0 +1,21 @@
+{
+  "name": "gtp_statically-strict-debugger",
+  "version": "1.0.0",
+  "description": "",
+  "main": "main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/RemedyOnline/gtp_smartHome_staticallyStrictDebugger.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "commonjs",
+  "bugs": {
+    "url": "https://github.com/RemedyOnline/gtp_smartHome_staticallyStrictDebugger/issues"
+  },
+  "homepage": "https://github.com/RemedyOnline/gtp_smartHome_staticallyStrictDebugger#readme"
+}
